# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Credit Broker 2.0",
        "description": "Документация\n * Во всех ответах приходит объект: `data`, кроме ошибок: В ошибках `401` имеется только поле `message`; в ошибке `422` имеются поля: `errors и message`.\n * Для авторизованных запросов необходимо вставить в HEADER `Authorization ` Bearer токен полученный при авторизации. Пример:\n *     Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n * Токен можно получить при входе в систему или регистрации. Сам токен будет расположен в ответе HEADER:\n *      X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n *      X-Access-Token-Expires-At: 2023-11-08 10:13:18\n *  Важно! На всех запросах, кроме DEV - необходимо отправлять в HEADER Platform-Code и Device-Id. Иначе будет приходить  `403 ошибка`\n *      Platform-Code: ios\n *      Device-Id: DFIE3232dD\n *\n * На всех запросах в HEADER ОБЯЗАТЕЛЬНО установить:\n *     Accept: application/json",
        "contact": {
            "email": "develop@tsay.uz"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API LOCAL SERVER"
        },
        {
            "url": "http://{tenant}.localhost",
            "description": "API LOCAL SERVER WITH TENANT",
            "variables": {
                "tenant": {
...
            }
...
    "paths": {
...
        "/v1/company/users": {
            "get": {
                "tags": [
                    "Пользователь компании"
                ],
                "summary": "Получение всех пользователей компании",
                "operationId": "62dd0a55e9514121666ed69f1512611d",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "post": {
                "tags": [
                    "Пользователь компании"
                ],
                "summary": "Добавление пользователя в компанию",
                "operationId": "c4d789a7d7f3c70fe5a140ed6f3c8522",
                "requestBody": {
                    "description": "Тело запроса на добавление пользователя в компанию",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "phone": "+998909101828"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/v1/company/users/{id}": {
            "get": {
                "tags": [
                    "Пользователь компании"
                ],
                "summary": "Получение информации о пользователе в компании",
                "operationId": "f1cc8b7f0b2ac66f6030d32ad05076ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "put": {
                "tags": [
                    "Пользователь компании"
                ],
                "summary": "Обновление информации о пользователе компании",
                "operationId": "0ba4700ced2f44cac10a4fc8c7724696",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Тело запроса для обновления данных пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "first_name": "Vasya",
                                "last_name": "Pupkin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Пользователь компании"
                ],
                "summary": "Удаление пользователя из компании",
                "operationId": "9ef0429bc6ff08908c047a7d81f78732",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "v1.Profile.UserResource": {
                "description": "Class UserResource.",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer",
                        "example": ""
                    },
                    "phone": {
                        "description": "Номер телефона пользователя",
                        "type": "string",
                        "example": ""
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": ""
                    },
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": ""
                    },
                    "is_active": {
                        "description": "Статус пользователя",
                        "type": "boolean",
                        "example": ""
                    },
                    "created_at": {
                        "description": "Дата создания сотрудника",
                        "type": "string",
                        "format": "date-time",
                        "example": ""
                    },
                    "updated_at": {
                        "description": "Дата обновления сотрудника",
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Пример: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjI0OT",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "Platform-Code": {
                "type": "apiKey",
                "description": "Код платформы",
                "name": "Platform-Code",
                "in": "header"
            },
            "Device-Id": {
                "type": "apiKey",
                "description": "ID устройства",
                "name": "Device-Id",
                "in": "header"
            },
            "Company-slug": {
                "type": "apiKey",
                "description": "Slug компании",
                "name": "Company",
                "in": "header"
            }
        }
    }
}